+++ src/Stack/Options/GhciParser.hs
@@ -1,13 +1,13 @@
 module Stack.Options.GhciParser where
 
-import           Data.Monoid.Extra
-import           Options.Applicative
-import           Options.Applicative.Args
-import           Options.Applicative.Builder.Extra
-import           Stack.Config                      (packagesParser)
-import           Stack.Ghci                        (GhciOpts (..))
-import           Stack.Options.BuildParser
-import           Stack.Types.Config
+import Data.Monoid.Extra
+import Data.Version (showVersion)
+import Options.Applicative
+import Options.Applicative.Args
+import Options.Applicative.Builder.Extra
+import Paths_stack as Meta
+import Stack.Config (packagesParser)
+import Stack.Ghci (GhciOpts (..))
 
 -- | Parser for GHCI options
 ghciOptsParser :: Parser GhciOpts
@@ -31,5 +31,12 @@ ghciOptsParser = GhciOpts
              <*> switch (long "load-local-deps" <> help "Load all local dependencies of your targets")
              <*> switch (long "skip-intermediate-deps" <> help "Skip loading intermediate target dependencies")
              <*> boolFlags True "package-hiding" "package hiding" idm
-             <*> buildOptsParser GhciBuildOpts Build
+             <*> many
+                   (textArgument
+                        (metavar "TARGET/FILE" <>
+                         help ("If none specified, use all local packages. " <>
+                               "See https://docs.haskellstack.org/en/v" <>
+                               showVersion Meta.version <>
+                               "/build_command/#target-syntax for details. " <>
+                               "If a path to a .hs or .lhs file is specified, it will be loaded.")))
              <*> switch (long "no-build" <> help "Don't build before launching GHCi (deprecated, should be unneeded)")