+++ ChangeLog.md
@@ -64,7 +64,7 @@ Other enhancements:
   on discussion in [#2647](https://github.com/commercialhaskell/stack/issues/2647).
 * The `--main-is` flag for GHCI now implies the TARGET, fixing
   [#1845](https://github.com/commercialhaskell/stack/issues/1845).
-* `stack ghci` will now use CLI `--ghc-options`.
+* `stack ghci` no longer takes all build options, as many weren't useful
   [#2199](https://github.com/commercialhaskell/stack/issues/2199)
 
 Bug fixes: